refactor(eventstore): split hook mappings and tests
