+++ packages/openclaw-nats-eventstore/src/hooks.ts
@@ -4,6 +4,13 @@ import type { PluginLogger } from "./nats-client.js";
 import type { NatsEventStoreConfig } from "./config.js";
 import type { ClawEvent, EventType, Visibility } from "./events.js";
 import { extractAgentId, buildSubject } from "./util.js";
+import {
+  HOOK_MAPPINGS,
+  EXTRA_EMITTERS,
+  type HookMapping,
+  type ExtraEmitter,
+  type EventTypeMapper,
+} from "./hook-mappings.js";
 
 type HookCtx = {
   agentId?: string;
@@ -89,6 +96,35 @@ function deriveEventId(
   return randomUUID();
 }
 
+function buildActor(agent: string, ctx: HookCtx): ClawEvent["actor"] {
+  return {
+    agentId: agent === "system" ? undefined : agent,
+    userId: firstString(ctx.senderId),
+    channel: firstString(ctx.channelId),
+  };
+}
+
+function buildScope(payload: Record<string, unknown>, ctx: HookCtx): ClawEvent["scope"] {
+  return {
+    sessionKey: firstString(ctx.sessionKey),
+    sessionId: firstString(ctx.sessionId),
+    runId: firstString(ctx.runId, payload.runId),
+    toolCallId: firstString(payload.toolCallId),
+    messageId: firstString(ctx.messageId, payload.messageId),
+    jobId: firstString(ctx.jobId),
+  };
+}
+
+function buildTrace(payload: Record<string, unknown>, ctx: HookCtx, trace: any): ClawEvent["trace"] {
+  return {
+    traceId: firstString(ctx.traceId, trace.traceId),
+    spanId: firstString(ctx.spanId, trace.spanId),
+    parentSpanId: firstString(ctx.parentSpanId, trace.parentSpanId),
+    causationId: firstString(payload.causationId),
+    correlationId: firstString(ctx.runId, ctx.sessionId, ctx.sessionKey),
+  };
+}
+
 function buildEnvelope(
   canonicalType: EventType,
   agent: string,
@@ -110,26 +146,9 @@ function buildEnvelope(
     legacyType: options.legacyType,
     schemaVersion: 1,
     source: { plugin: "nats-eventstore" },
-    actor: {
-      agentId: agent === "system" ? undefined : agent,
-      userId: firstString(ctx.senderId),
-      channel: firstString(ctx.channelId),
-    },
-    scope: {
-      sessionKey: firstString(ctx.sessionKey),
-      sessionId: firstString(ctx.sessionId),
-      runId: firstString(ctx.runId, payload.runId),
-      toolCallId: firstString(payload.toolCallId),
-      messageId: firstString(ctx.messageId, payload.messageId),
-      jobId: firstString(ctx.jobId),
-    },
-    trace: {
-      traceId: firstString(ctx.traceId, trace.traceId),
-      spanId: firstString(ctx.spanId, trace.spanId),
-      parentSpanId: firstString(ctx.parentSpanId, trace.parentSpanId),
-      causationId: firstString(payload.causationId),
-      correlationId: firstString(ctx.runId, ctx.sessionId, ctx.sessionKey),
-    },
+    actor: buildActor(agent, ctx),
+    scope: buildScope(payload, ctx),
+    trace: buildTrace(payload, ctx, trace),
     visibility: options.visibility ?? "internal",
     redaction: options.redaction,
     payload,
@@ -158,223 +177,78 @@ function publish(
   });
 }
 
-// ── Hook-to-event mapping table ──
+function resolveEventType(mapper: EventTypeMapper, event: any, ctx: any): EventType {
+  return typeof mapper === "function" ? mapper(event, ctx) : mapper;
+}
 
-type PayloadMapper = (event: any, ctx: any) => Record<string, unknown>;
-type EventTypeMapper = EventType | ((event: any, ctx: any) => EventType);
+function emitExtras(
+  extras: ExtraEmitter[],
+  event: any,
+  ctx: any,
+  pub: (type: EventType, ctxPayload: HookCtx, payload: Record<string, unknown>, options?: PublishOptions) => void
+): void {
+  for (const extra of extras) {
+    if (extra.condition(event)) {
+      const extraType = resolveEventType(extra.eventType, event, ctx);
+      pub(extraType, ctx, extra.mapper(event, ctx), {
+        legacyType: extra.legacyType,
+        visibility: extra.visibility,
+        redaction: extra.redaction,
+        originalEvent: event,
+      });
+    }
+  }
+}
 
-type HookMapping = {
-  hookName: string;
-  eventType: EventTypeMapper;
-  legacyType?: EventType;
-  visibility?: Visibility;
-  redaction?: ClawEvent["redaction"];
-  mapper: PayloadMapper;
-  /** If true, uses "system" as agent/session (gateway hooks). */
-  systemEvent?: boolean;
-};
+function publishHookResult(
+  getClient: () => NatsClient | null,
+  config: NatsEventStoreConfig,
+  logger: PluginLogger
+) {
+  return (
+    type: EventType,
+    ctxPayload: HookCtx,
+    payload: Record<string, unknown>,
+    options: PublishOptions = {},
+  ): void => {
+    const agent = extractAgentId(ctxPayload);
+    const session = ctxPayload.sessionKey ?? ctxPayload.sessionId ?? "unknown";
+    publish(getClient, config, type, agent, session, payload, logger, { ...options, ctx: ctxPayload });
+  };
+}
 
-/** Additional events to emit from the same hook (e.g. run.error from agent_end). */
-type ExtraEmitter = {
-  hookName: string;
-  eventType: EventTypeMapper;
-  legacyType?: EventType;
-  visibility?: Visibility;
-  redaction?: ClawEvent["redaction"];
-  condition: (event: any) => boolean;
-  mapper: PayloadMapper;
-};
+function handleHookEvent(
+  mapping: HookMapping,
+  extras: ExtraEmitter[],
+  event: any,
+  ctx: any,
+  config: NatsEventStoreConfig,
+  getClient: () => NatsClient | null,
+  logger: PluginLogger,
+): void {
+  try {
+    const pub = publishHookResult(getClient, config, logger);
+    const payload = mapping.mapper(event, ctx);
+    const eventType = resolveEventType(mapping.eventType, event, ctx);
+    const options: PublishOptions = {
+      legacyType: mapping.legacyType,
+      visibility: mapping.visibility,
+      redaction: mapping.redaction,
+      ctx,
+      originalEvent: event,
+    };
 
-const HOOK_MAPPINGS: HookMapping[] = [
-  {
-    hookName: "message_received",
-    eventType: "message.in.received",
-    legacyType: "msg.in",
-    visibility: "confidential",
-    mapper: (event, ctx) => ({
-      from: event.from,
-      content: event.content,
-      timestamp: event.timestamp,
-      channel: ctx?.channelId,
-      metadata: event.metadata,
-    }),
-  },
-  {
-    hookName: "message_sending",
-    eventType: "message.out.sending",
-    legacyType: "msg.sending",
-    visibility: "confidential",
-    mapper: (event, ctx) => ({
-      to: event.to,
-      content: event.content,
-      channel: ctx?.channelId,
-    }),
-  },
-  {
-    hookName: "message_sent",
-    eventType: "message.out.sent",
-    legacyType: "msg.out",
-    visibility: "confidential",
-    mapper: (event, ctx) => ({
-      to: event.to,
-      content: event.content,
-      success: event.success,
-      error: event.error,
-      channel: ctx?.channelId,
-    }),
-  },
-  {
-    hookName: "before_tool_call",
-    eventType: "tool.call.requested",
-    legacyType: "tool.call",
-    visibility: "confidential",
-    mapper: (event) => ({
-      toolName: event.toolName,
-      params: event.params,
-    }),
-  },
-  {
-    hookName: "after_tool_call",
-    eventType: (event) => event.error ? "tool.call.failed" : "tool.call.executed",
-    legacyType: "tool.result",
-    visibility: "confidential",
-    mapper: (event) => ({
-      toolName: event.toolName,
-      params: event.params,
-      result: event.result,
-      error: event.error,
-      durationMs: event.durationMs,
-    }),
-  },
-  {
-    hookName: "before_agent_start",
-    eventType: "run.started",
-    legacyType: "run.start",
-    visibility: "confidential",
-    mapper: (event) => ({ prompt: event.prompt }),
-  },
-  {
-    hookName: "agent_end",
-    eventType: "run.ended",
-    legacyType: "run.end",
-    mapper: (event) => ({
-      success: event.success,
-      error: event.error,
-      durationMs: event.durationMs,
-      messageCount: event.messages?.length ?? 0,
-    }),
-  },
-  {
-    hookName: "llm_input",
-    eventType: "model.input.observed",
-    legacyType: "llm.input",
-    redaction: { applied: true, omittedFields: ["systemPrompt", "prompt", "historyMessages"] },
-    mapper: (event) => ({
-      runId: event.runId,
-      sessionId: event.sessionId,
-      provider: event.provider,
-      model: event.model,
-      systemPromptLength: event.systemPrompt?.length ?? 0,
-      promptLength: event.prompt?.length ?? 0,
-      historyMessageCount: event.historyMessages?.length ?? 0,
-      imagesCount: event.imagesCount ?? 0,
-    }),
-  },
-  {
-    hookName: "llm_output",
-    eventType: "model.output.observed",
-    legacyType: "llm.output",
-    redaction: { applied: true, omittedFields: ["assistantTexts"] },
-    mapper: (event) => {
-      const texts = event.assistantTexts ?? [];
-      return {
-        runId: event.runId,
-        sessionId: event.sessionId,
-        provider: event.provider,
-        model: event.model,
-        assistantTextCount: texts.length,
-        assistantTextTotalLength: texts.reduce(
-          (s: number, t: string) => s + (t?.length ?? 0),
-          0,
-        ),
-        usage: event.usage,
-      };
-    },
-  },
-  {
-    hookName: "before_compaction",
-    eventType: "session.compaction.started",
-    legacyType: "session.compaction_start",
-    mapper: (event) => ({
-      messageCount: event.messageCount,
-      compactingCount: event.compactingCount,
-      tokenCount: event.tokenCount,
-    }),
-  },
-  {
-    hookName: "after_compaction",
-    eventType: "session.compaction.ended",
-    legacyType: "session.compaction_end",
-    mapper: (event) => ({
-      messageCount: event.messageCount,
-      compactedCount: event.compactedCount,
-      tokenCount: event.tokenCount,
-    }),
-  },
-  {
-    hookName: "before_reset",
-    eventType: "session.reset",
-    mapper: (event) => ({ reason: event.reason }),
-  },
-  {
-    hookName: "session_start",
-    eventType: "session.started",
-    legacyType: "session.start",
-    mapper: (event) => ({
-      sessionId: event.sessionId,
-      resumedFrom: event.resumedFrom,
-    }),
-  },
-  {
-    hookName: "session_end",
-    eventType: "session.ended",
-    legacyType: "session.end",
-    mapper: (event) => ({
-      sessionId: event.sessionId,
-      messageCount: event.messageCount,
-      durationMs: event.durationMs,
-    }),
-  },
-  {
-    hookName: "gateway_start",
-    eventType: "gateway.started",
-    legacyType: "gateway.start",
-    mapper: (event) => ({ port: event.port }),
-    systemEvent: true,
-  },
-  {
-    hookName: "gateway_stop",
-    eventType: "gateway.stopped",
-    legacyType: "gateway.stop",
-    mapper: (event) => ({ reason: event.reason }),
-    systemEvent: true,
-  },
-];
+    if (mapping.systemEvent) {
+      publish(getClient, config, eventType, "system", "system", payload, logger, options);
+    } else {
+      pub(eventType, ctx, payload, options);
+    }
 
-/** Extra events emitted from existing hooks (backward compatibility). */
-const EXTRA_EMITTERS: ExtraEmitter[] = [
-  {
-    hookName: "agent_end",
-    eventType: "run.failed",
-    legacyType: "run.error",
-    condition: (event) => !event.success,
-    mapper: (event) => ({
-      success: false,
-      error: event.error,
-      durationMs: event.durationMs,
-    }),
-  },
-];
+    emitExtras(extras, event, ctx, pub);
+  } catch (err) {
+    logger.warn(`[nats-eventstore] Hook ${mapping.hookName} error: ${err}`);
+  }
+}
 
 /**
  * Register all event hook handlers on the plugin API.
@@ -385,24 +259,6 @@ export function registerEventHooks(
   config: NatsEventStoreConfig,
   getClient: () => NatsClient | null,
 ): void {
-  const logger = api.logger;
-
-  // Helper: publish with agent context extraction
-  const pub = (
-    type: EventType,
-    ctx: HookCtx,
-    payload: Record<string, unknown>,
-    options: PublishOptions = {},
-  ) => {
-    const agent = extractAgentId(ctx);
-    const session = ctx.sessionKey ?? ctx.sessionId ?? "unknown";
-    publish(getClient, config, type, agent, session, payload, logger, { ...options, ctx });
-  };
-
-  const resolveEventType = (mapper: EventTypeMapper, event: any, ctx: any): EventType =>
-    typeof mapper === "function" ? mapper(event, ctx) : mapper;
-
-  // Build lookup for extra emitters per hook
   const extrasByHook = new Map<string, ExtraEmitter[]>();
   for (const extra of EXTRA_EMITTERS) {
     const list = extrasByHook.get(extra.hookName) ?? [];
@@ -410,46 +266,11 @@ export function registerEventHooks(
     extrasByHook.set(extra.hookName, list);
   }
 
-  // Register each hook from the mapping table
   for (const mapping of HOOK_MAPPINGS) {
     if (!shouldPublish(mapping.hookName, config)) continue;
-
     const extras = extrasByHook.get(mapping.hookName) ?? [];
-
     api.on(mapping.hookName, (event: any, ctx: any) => {
-      try {
-        const payload = mapping.mapper(event, ctx);
-
-        const eventType = resolveEventType(mapping.eventType, event, ctx);
-        const options: PublishOptions = {
-          legacyType: mapping.legacyType,
-          visibility: mapping.visibility,
-          redaction: mapping.redaction,
-          ctx,
-          originalEvent: event,
-        };
-
-        if (mapping.systemEvent) {
-          publish(getClient, config, eventType, "system", "system", payload, logger, options);
-        } else {
-          pub(eventType, ctx, payload, options);
-        }
-
-        // Emit any extra events for this hook
-        for (const extra of extras) {
-          if (extra.condition(event)) {
-            const extraType = resolveEventType(extra.eventType, event, ctx);
-            pub(extraType, ctx, extra.mapper(event, ctx), {
-              legacyType: extra.legacyType,
-              visibility: extra.visibility,
-              redaction: extra.redaction,
-              originalEvent: event,
-            });
-          }
-        }
-      } catch (err) {
-        logger.warn(`[nats-eventstore] Hook ${mapping.hookName} error: ${err}`);
-      }
+      handleHookEvent(mapping, extras, event, ctx, config, getClient, api.logger);
     });
   }
 }
